function_disassembly() account for edge case
When attempting to find the gdb.Block() of a function, we use
gdb.block_for_pc() of the function address. This usually either raises a
RuntimeError or returns None if there is a problem (what happens depends
on what the problem is).
Sometimes (I guess when this comes from a file compiled without
debugging information but compiled as part of a program that has
debugging information) we get a file-level block instead of a
function block. This is just as useless to us as no block at all, so
check for that and fall back to the alternate method.