Never manually parse the argument
Before we manually split our argument on `{}`, but now we always just
pass it to string.format(). This means that everything is consistent, so
we can use {.t} or {.v} anywhere. Unfortunately it also means we have to
use {0} everywhere with more than one use, but that's a worthwhile
price.
Also add some more tests.