Add a std::list walker for libstdc++
I want walkers in C++ (seeing as I now have to learn it ...).
As of yet I don't know aproximately how stable the implementations will
be. There's nothing enforcing the implementation of std::list to stay
the same, but the tool will be helpful just as long as the
implementation doesn't change *too* often.