helpers.file_symbols() add another special case
As far as I know there's no way to only catch errors that I care about
reliably, so I'm stuck with this.
The error messages aren't standardised and there is no class hierarchy
from gdb.error that marks out the file not existing from any other
errors that might happen.

It may be that there is a specification here that I'm not aware of ...
if so I'll find out in the end.