Get autoimports through the PathFinder
Before this change we added a special finder to sys.meta_path, this
was called for every failed import, and always felt like overkill.
Now we put a callable on sys.path_hooks(), which is only called when a
new path is checked for importing, and the results are cached.

This is a less invasive way to get the same effect.