Add programmatic interface to walker creation
Close #3

This is mainly because it "feels right" to have a programmatic
interface.
Without one any walkers that want to manage their own pipeline have to
manipulate strings which is not nice (as SQL prooves).

As of yet I haven't made any walkers sufficiently complex to make this a
problem, but I feel it's something better done before the source base
grows enough to make it a big chore.