Give called-functions a 'unique' option
This allows limiting the returned functions to unique ones.
By contrast, the default operation limits itself to unique stacks.

The first is useful when looking for the set of functions this one
calls, the second is useful when wanting to understand how this function
can get to each called function in the tree.