Add boost_interprocess walkers
This module is for walking across containers defined in the
boost::interprocess library.
These are the same containers as defined in the boost::containers
library.
This is not put in autoloads/ as the libraries are header-only, which
means there is no shared object load to hook into.

Hence the walkers have to be loaded with
`python import boost_interprocess`