Fixed `array auto; ...` bug introduced in b91aff6
When `array auto; ...` is used in subsequent steps of a pipeline, it
records the fact that it doesn't know the type of elements it's using by
setting `self.typename` to `None`.

When we switched from giving the array walker the type of pointers to
elements in the array to giving the walker the type of elements in the
array the `__init__` function was updated to modify `self.typename` by
adding a `*` to the end.
This didn't take into account that `self.typename` could be `None`.

This commit adds a conditional to that operation.
I'm a little annoyed that this problem was caught by my test suite.
That means I didn't bother to run the testsuite against my change, or at
least didn't look into the failures.