Add the if-nth walker
This returns the element at an offset N to the element that matches a
criteria.

I'd put off defining this walker since it seems like a hack just for the
problem I happen to be facing right then, but I've wanted it enough
times now to put it in anyway.