Working towards only using address gdb.Values
This is a bit of a compromise.
It means we don't get any extra use of the pretty_printer.children()
methods out there which provide gdb.Values without addresses.

However, it means we can use other languages as long as they can cast a
raw address to a value.

It also means we get a lot more type parsing robustness.