Add a "machine interface" fallback for `search_symbol_wrapper'
This was surprisingly easy -- I originally tried to use `gdb.execute'
with a command of `interpreter-exec mi ...' to capture the output, then
realised the output from that would not get captured.

Later went through a whole arc of using `new-ui', with a PTY I just
opened and writing a parser combinator to get GDB output (which took a
whole night).

Then I realised that had a real problem that GDB wouldn't answer the
request sent on the PTY until the python stopped and let GDB's event
queue pick it up.

While searching for that I found the `gdb.execute_mi' command, which
obviated the need for writing that parser combinator at all.

A great lesson about "look for an easier approach before starting".
But also quite a good lesson in writing a parser combinator (so not a
total loss).