autoimport account for `rr` using different filename
Noticing that I'm not getting my walkers because `rr` is using a
different objfile name to the original object file.

I have the feeling this worked before, but this is a fix that I doubt is
going to have problems anywhere else and gets things working, so may as
well just make the change.