Account for embedded case in GCC vec walker
There are two different memory allocation mechanisms for the GCC vec
structure.  One is to have a vector prefix allocated directly in front
of the contained data.  The other is to have a pointer pointing to both
a prefix and a header.

I'd not noticed the embedded version when originally writing this.  This
commit accounts for the embedded version by checking which is happening.

I also think I was using a problematic approach to identify the type of
the vector.  Adjust to attempt to query the type of the pointed to
gdb.Type rather than the type of the pointer.  This seems to catch more
cases.